@@ -2,7 +2,7 @@ class SiteBaseController  < BaseController
 
 
   ### custom for application
-  before_filter :my_before_page_render
+  before_action :my_before_page_render
 
 
   def my_before_page_render
