@@ -1,8 +1,8 @@
 ﻿//= require jquery2
 //= require jquery_ujs
 
-//= require popper
-//= require bootstrap
+//= require popper.js/dist/umd/popper.js
+//= require bootstrap/dist/js/bootstrap
 
 //= require bootstrap3-typeahead.min
 //= require bootstrap-autocomplete-input.min
