@@ -17,7 +17,7 @@ module Myapp
     # config.time_zone = 'Central Time (US & Canada)'
 
     # Do not swallow errors in after_commit/after_rollback callbacks.
-    config.active_record.raise_in_transactional_callbacks = true
+    #config.active_record.raise_in_transactional_callbacks = true
 
 
     #
